Move types around, use same type for h/v center, add to prelude
