+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Unreleased
 ## Added:
 
  * Support for vertical text alignment.
+ * Added alignment types to `prelude`.
  * Support for zero-width space character (`\u{200B}`).
  * Support for nonbreaking space character (`\u{A0}`).
  * Added optimized measurement function that do not expect carriage returns.
@@ -15,10 +16,6 @@ Unreleased
 ## Fixed:
 
  * Fixed an issue where height measurement unexpectedly carried a space that is consumed during drawing.
-
-## Changed:
-
- * **breaking** Moved horizontal text alignment types to `alignment::horizontal`.
 
 0.1.0 (2020-07-31)
 ==================
