+++ examples/center.rs
@@ -1,10 +1,8 @@
+use embedded_graphics::{fonts::Font6x8, pixelcolor::BinaryColor, prelude::*};
 use embedded_graphics_simulator::{
     BinaryColorTheme, OutputSettingsBuilder, SimulatorDisplay, Window,
 };
-
-use embedded_graphics::{fonts::Font6x8, pixelcolor::BinaryColor, prelude::*};
-
-use embedded_text::{alignment::horizontal::CenterAligned, prelude::*};
+use embedded_text::prelude::*;
 
 fn main() -> Result<(), core::convert::Infallible> {
     let text = "Hello, World!\nLorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book.";
