+++ examples/color.rs
@@ -1,8 +1,6 @@
-use embedded_graphics_simulator::{OutputSettingsBuilder, SimulatorDisplay, Window};
-
 use embedded_graphics::{fonts::Font6x8, pixelcolor::Rgb565, prelude::*};
-
-use embedded_text::{alignment::horizontal::Justified, prelude::*};
+use embedded_graphics_simulator::{OutputSettingsBuilder, SimulatorDisplay, Window};
+use embedded_text::prelude::*;
 
 fn main() -> Result<(), core::convert::Infallible> {
     let mut display: SimulatorDisplay<Rgb565> = SimulatorDisplay::new(Size::new(129, 129));
