+++ examples/interactive.rs
@@ -12,14 +12,9 @@ use embedded_graphics::{
     prelude::*,
     style::PrimitiveStyle,
 };
+use embedded_text::prelude::*;
 use sdl2::keyboard::Keycode;
 use std::{thread, time::Duration};
-
-use embedded_text::{
-    alignment::{horizontal::*, vertical::top::Top},
-    prelude::*,
-    style::StyledTextBox,
-};
 
 enum ProcessedEvent {
     Nothing,
@@ -66,7 +61,7 @@ impl ProcessedEvent {
 fn demo_loop<A>(window: &mut Window, bounds: &mut Rectangle, alignment: A) -> bool
 where
     A: HorizontalTextAlignment,
-    for<'a> &'a StyledTextBox<'a, BinaryColor, Font6x8, A, Top>: Drawable<BinaryColor>,
+    for<'a> &'a StyledTextBox<'a, BinaryColor, Font6x8, A, TopAligned>: Drawable<BinaryColor>,
 {
     loop {
         let mut display: SimulatorDisplay<BinaryColor> = SimulatorDisplay::new(Size::new(255, 255));
