+++ src/alignment/justified.rs
@@ -1,10 +1,10 @@
-//! Left aligned text.
+//! Fully justified text.
 use crate::{
-    alignment::{horizontal::HorizontalTextAlignment, vertical::VerticalTextAlignment},
+    alignment::{HorizontalTextAlignment, VerticalTextAlignment},
     parser::Token,
     rendering::{
         cursor::Cursor,
-        line::{StyledLineIterator, UniformSpaceConfig},
+        line::{SpaceConfig, StyledLineIterator},
         StateFactory, StyledTextBoxIterator,
     },
     style::StyledTextBox,
@@ -12,15 +12,63 @@ use crate::{
 };
 use embedded_graphics::{drawable::Pixel, fonts::Font, pixelcolor::PixelColor};
 
-/// Marks text to be rendered left aligned.
+/// Marks text to be rendered fully justified.
 #[derive(Copy, Clone, Debug)]
-pub struct LeftAligned;
-impl HorizontalTextAlignment for LeftAligned {
-    const STARTING_SPACES: bool = true;
-    const ENDING_SPACES: bool = true;
+pub struct Justified;
+impl HorizontalTextAlignment for Justified {
+    const STARTING_SPACES: bool = false;
+    const ENDING_SPACES: bool = false;
 }
 
-/// State variable used by the left aligned text renderer.
+/// Internal state information used to store width of whitespace characters when rendering fully
+/// justified text.
+///
+/// The fully justified renderer works by calculating the width of whitespace characters for the
+/// current line. Due to integer arithmetic, there can be remainder pixels when a single space
+/// width is used. This struct stores two width values so the whole line will always (at least if
+/// there's a space in the line) take up all available space.
+#[derive(Copy, Clone, Debug)]
+pub struct JustifiedSpaceConfig {
+    /// The width of the whitespace characters.
+    space_width: u32,
+
+    /// Stores how many characters are rendered using the space_width width. This field changes
+    /// during rendering.
+    space_count: u32,
+}
+
+impl JustifiedSpaceConfig {
+    #[inline]
+    #[must_use]
+    fn default<F: Font>() -> Self {
+        JustifiedSpaceConfig::new(F::total_char_width(' '), 0)
+    }
+
+    #[inline]
+    #[must_use]
+    fn new(space_width: u32, extra_pixel_count: u32) -> Self {
+        JustifiedSpaceConfig {
+            space_width,
+            space_count: extra_pixel_count,
+        }
+    }
+}
+
+impl SpaceConfig for JustifiedSpaceConfig {
+    #[inline]
+    fn peek_next_width(&self, whitespace_count: u32) -> u32 {
+        whitespace_count * self.space_width + self.space_count.min(whitespace_count)
+    }
+
+    #[inline]
+    fn consume(&mut self, n: u32) -> u32 {
+        let w = self.peek_next_width(n);
+        self.space_count = self.space_count.saturating_sub(n);
+        w
+    }
+}
+
+/// State variable used by the fully justified text renderer.
 #[derive(Debug)]
 pub enum State<'a, C, F>
 where
@@ -31,10 +79,10 @@ where
     NextLine(Option<Token<'a>>, Cursor<F>),
 
     /// Renders the processed line.
-    DrawLine(StyledLineIterator<'a, C, F, UniformSpaceConfig, LeftAligned>),
+    DrawLine(StyledLineIterator<'a, C, F, JustifiedSpaceConfig, Justified>),
 }
 
-impl<'a, C, F, V> StateFactory<F> for StyledTextBox<'a, C, F, LeftAligned, V>
+impl<'a, C, F, V> StateFactory<F> for StyledTextBox<'a, C, F, Justified, V>
 where
     C: PixelColor,
     F: Font + Copy,
@@ -49,7 +97,7 @@ where
     }
 }
 
-impl<C, F, V> Iterator for StyledTextBoxIterator<'_, C, F, LeftAligned, V>
+impl<C, F, V> Iterator for StyledTextBoxIterator<'_, C, F, Justified, V>
 where
     C: PixelColor,
     F: Font + Copy,
@@ -70,12 +118,29 @@ where
                         break None;
                     }
 
+                    let max_line_width = cursor.line_width();
+                    let (width, total_whitespace_count, t) = self.style.measure_line(
+                        &mut self.parser.clone(),
+                        carried_token.clone(),
+                        max_line_width,
+                    );
+
+                    let space = max_line_width
+                        - (width - total_whitespace_count * F::total_char_width(' '));
+                    let stretch_line = t.is_some() && t != Some(Token::NewLine);
+
+                    let space_info = if stretch_line && total_whitespace_count != 0 {
+                        let space_width = space / total_whitespace_count;
+                        let extra_pixels = space % total_whitespace_count;
+                        JustifiedSpaceConfig::new(space_width, extra_pixels)
+                    } else {
+                        JustifiedSpaceConfig::default::<F>()
+                    };
+
                     self.state = State::DrawLine(StyledLineIterator::new(
                         self.parser.clone(),
                         *cursor,
-                        UniformSpaceConfig {
-                            space_width: F::total_char_width(' '),
-                        },
+                        space_info,
                         self.style.text_style,
                         carried_token.clone(),
                     ));
@@ -109,7 +174,7 @@ where
 
                     self.state = State::NextLine(carried_token, line_iterator.cursor);
                 }
-            };
+            }
         }
     }
 }
@@ -121,13 +186,13 @@ mod test {
         primitives::Rectangle,
     };
 
-    use crate::{alignment::horizontal::LeftAligned, style::TextBoxStyleBuilder, TextBox};
+    use crate::{alignment::Justified, style::TextBoxStyleBuilder, TextBox};
 
     #[test]
     fn simple_render() {
         let mut display = MockDisplay::new();
         let style = TextBoxStyleBuilder::new(Font6x8)
-            .alignment(LeftAligned)
+            .alignment(Justified)
             .text_color(BinaryColor::On)
             .background_color(BinaryColor::Off)
             .build();
@@ -156,7 +221,7 @@ mod test {
     fn simple_render_cr() {
         let mut display = MockDisplay::new();
         let style = TextBoxStyleBuilder::new(Font6x8)
-            .alignment(LeftAligned)
+            .alignment(Justified)
             .text_color(BinaryColor::On)
             .build();
 
@@ -180,10 +245,42 @@ mod test {
     }
 
     #[test]
+    fn wrapping_when_space_is_less_than_space_character() {
+        let mut display = MockDisplay::new();
+        let style = TextBoxStyleBuilder::new(Font6x8)
+            .alignment(Justified)
+            .text_color(BinaryColor::On)
+            .background_color(BinaryColor::Off)
+            .build();
+
+        TextBox::new(
+            "A word",
+            Rectangle::new(Point::zero(), Point::new(6 * 5 - 1, 7)),
+        )
+        .into_styled(style)
+        .draw(&mut display)
+        .unwrap();
+
+        assert_eq!(
+            display,
+            MockDisplay::from_pattern(&[
+                ".###..            ",
+                "#...#.            ",
+                "#...#.            ",
+                "#####.            ",
+                "#...#.            ",
+                "#...#.            ",
+                "#...#.            ",
+                "......            ",
+            ])
+        );
+    }
+
+    #[test]
     fn simple_word_wrapping() {
         let mut display = MockDisplay::new();
         let style = TextBoxStyleBuilder::new(Font6x8)
-            .alignment(LeftAligned)
+            .alignment(Justified)
             .text_color(BinaryColor::On)
             .background_color(BinaryColor::Off)
             .build();
@@ -199,14 +296,14 @@ mod test {
         assert_eq!(
             display,
             MockDisplay::from_pattern(&[
-                "......................#.......                  ",
-                "......................#.......                  ",
-                "#...#..###..#.##...##.#.......                  ",
-                "#...#.#...#.##..#.#..##.......                  ",
-                "#.#.#.#...#.#.....#...#.......                  ",
-                "#.#.#.#...#.#.....#...#.......                  ",
-                ".#.#...###..#......####.......                  ",
-                "..............................                  ",
+                "......................#.                        ",
+                "......................#.                        ",
+                "#...#..###..#.##...##.#.                        ",
+                "#...#.#...#.##..#.#..##.                        ",
+                "#.#.#.#...#.#.....#...#.                        ",
+                "#.#.#.#...#.#.....#...#.                        ",
+                ".#.#...###..#......####.                        ",
+                "........................                        ",
                 "................................#...............",
                 "................................................",
                 "#...#.#.##...###..####..####...##...#.##...####.",
@@ -220,17 +317,17 @@ mod test {
     }
 
     #[test]
-    fn simple_word_wrapping_by_space() {
+    fn justified_alignment() {
         let mut display = MockDisplay::new();
         let style = TextBoxStyleBuilder::new(Font6x8)
-            .alignment(LeftAligned)
+            .alignment(Justified)
             .text_color(BinaryColor::On)
             .background_color(BinaryColor::Off)
             .build();
 
         TextBox::new(
-            "wrapping word",
-            Rectangle::new(Point::zero(), Point::new(47, 15)),
+            "word and other word last line",
+            Rectangle::new(Point::zero(), Point::new(60, 23)),
         )
         .into_styled(style)
         .draw(&mut display)
@@ -239,62 +336,30 @@ mod test {
         assert_eq!(
             display,
             MockDisplay::from_pattern(&[
-                "................................#...............",
-                "................................................",
-                "#...#.#.##...###..####..####...##...#.##...####.",
-                "#...#.##..#.....#.#...#.#...#...#...##..#.#...#.",
-                "#.#.#.#......####.#...#.#...#...#...#...#.#...#.",
-                "#.#.#.#.....#...#.####..####....#...#...#..####.",
-                ".#.#..#......####.#.....#......###..#...#.....#.",
-                "..................#.....#..................###..",
-                "......................#.                        ",
-                "......................#.                        ",
-                "#...#..###..#.##...##.#.                        ",
-                "#...#.#...#.##..#.#..##.                        ",
-                "#.#.#.#...#.#.....#...#.                        ",
-                "#.#.#.#...#.#.....#...#.                        ",
-                ".#.#...###..#......####.                        ",
-                "........................                        ",
-            ])
-        );
-    }
-
-    #[test]
-    fn whitespace_word_wrapping() {
-        let mut display = MockDisplay::new();
-        let style = TextBoxStyleBuilder::new(Font6x8)
-            .alignment(LeftAligned)
-            .text_color(BinaryColor::On)
-            .background_color(BinaryColor::Off)
-            .build();
-
-        TextBox::new(
-            "word  wrap",
-            Rectangle::new(Point::zero(), Point::new(30, 15)),
-        )
-        .into_styled(style)
-        .draw(&mut display)
-        .unwrap();
-
-        assert_eq!(
-            display,
-            MockDisplay::from_pattern(&[
-                "......................#.......",
-                "......................#.......",
-                "#...#..###..#.##...##.#.......",
-                "#...#.#...#.##..#.#..##.......",
-                "#.#.#.#...#.#.....#...#.......",
-                "#.#.#.#...#.#.....#...#.......",
-                ".#.#...###..#......####.......",
-                "..............................",
-                "..............................",
-                "..............................",
-                "......#...#.#.##...###..####..",
-                "......#...#.##..#.....#.#...#.",
-                "......#.#.#.#......####.#...#.",
-                "......#.#.#.#.....#...#.####..",
-                ".......#.#..#......####.#.....",
-                "........................#....."
+                "......................#....................................#.",
+                "......................#....................................#.",
+                "#...#..###..#.##...##.#.....................###..#.##...##.#.",
+                "#...#.#...#.##..#.#..##........................#.##..#.#..##.",
+                "#.#.#.#...#.#.....#...#.....................####.#...#.#...#.",
+                "#.#.#.#...#.#.....#...#....................#...#.#...#.#...#.",
+                ".#.#...###..#......####.....................####.#...#..####.",
+                ".............................................................",
+                ".......#....#..............................................#.",
+                ".......#....#..............................................#.",
+                ".###..###...#.##...###..#.##.........#...#..###..#.##...##.#.",
+                "#...#..#....##..#.#...#.##..#........#...#.#...#.##..#.#..##.",
+                "#...#..#....#...#.#####.#............#.#.#.#...#.#.....#...#.",
+                "#...#..#..#.#...#.#.....#............#.#.#.#...#.#.....#...#.",
+                ".###....##..#...#..###..#.............#.#...###..#......####.",
+                ".............................................................",
+                ".##................#...........##.....#...............       ",
+                "..#................#............#.....................       ",
+                "..#....###...####.###...........#....##...#.##...###..       ",
+                "..#.......#.#......#............#.....#...##..#.#...#.       ",
+                "..#....####..###...#............#.....#...#...#.#####.       ",
+                "..#...#...#.....#..#..#.........#.....#...#...#.#.....       ",
+                ".###...####.####....##.........###...###..#...#..###..       ",
+                "......................................................       "
             ])
         );
     }
@@ -303,7 +368,7 @@ mod test {
     fn word_longer_than_line_wraps_word() {
         let mut display = MockDisplay::new();
         let style = TextBoxStyleBuilder::new(Font6x8)
-            .alignment(LeftAligned)
+            .alignment(Justified)
             .text_color(BinaryColor::On)
             .background_color(BinaryColor::Off)
             .build();
@@ -319,14 +384,14 @@ mod test {
         assert_eq!(
             display,
             MockDisplay::from_pattern(&[
-                "......................#.......                        ",
-                "......................#.......                        ",
-                "#...#..###..#.##...##.#.......                        ",
-                "#...#.#...#.##..#.#..##.......                        ",
-                "#.#.#.#...#.#.....#...#.......                        ",
-                "#.#.#.#...#.#.....#...#.......                        ",
-                ".#.#...###..#......####.......                        ",
-                "..............................                        ",
+                "......................#.                              ",
+                "......................#.                              ",
+                "#...#..###..#.##...##.#.                              ",
+                "#...#.#...#.##..#.#..##.                              ",
+                "#.#.#.#...#.#.....#...#.                              ",
+                "#.#.#.#...#.#.....#...#.                              ",
+                ".#.#...###..#......####.                              ",
+                "........................                              ",
                 "...........................................##....##...",
                 "............................................#.....#...",
                 ".####..###..##.#...###..#.##...###...###....#.....#...",
@@ -351,7 +416,7 @@ mod test {
     fn first_word_longer_than_line_wraps_word() {
         let mut display = MockDisplay::new();
         let style = TextBoxStyleBuilder::new(Font6x8)
-            .alignment(LeftAligned)
+            .alignment(Justified)
             .text_color(BinaryColor::On)
             .background_color(BinaryColor::Off)
             .build();
